move set_num_periods out
