+++ src/ibex_bluesky_core/plans/reflectometry/crisp.py
@@ -25,11 +25,8 @@ from ibex_bluesky_core.devices.simpledae.controllers import (
 from ibex_bluesky_core.devices.simpledae.reducers import MonitorNormalizer
 from ibex_bluesky_core.devices.simpledae.waiters import GoodFramesWaiter, PeriodGoodFramesWaiter
 from ibex_bluesky_core.plan_stubs import call_qt_aware
-from ibex_bluesky_core.plans.reflectometry.utils import (
-    centred_pixel,
-    refl_parameter,
-    set_num_periods,
-)
+from ibex_bluesky_core.plans import set_num_periods
+from ibex_bluesky_core.plans.reflectometry import centred_pixel, refl_parameter
 from ibex_bluesky_core.run_engine import get_run_engine
 
 matplotlib.rcParams["figure.autolayout"] = True
